refactor agent task components, turn on autorefresh by default (#130)
* refactor agent task components, turn on autorefresh by default

* Apply suggestions from code review

* fix initializing of interval

* Update CHANGELOG.md
